requirements.txt
flask>=2.0.0
kiteconnect>=4.1.0
pandas>=1.3.0
yfinance>=0.1.70
python-telegram-bot>=13.7
razorpay>=1.3.0
websocket-client>=1.2.1
APScheduler>=3.8.1
requests>=2.26.0
python-dotenv>=0.19.0
gunicorn>=20.1.0  # For production deployment